Adds unit tests and fixes error callback.
According to the Node.js documentation, APIs should be
either asynchronous or synchronous -- not both, which
was the case in this library. Previously, if parameters
was not defined, it would synchronously callback and
could not be depended on. Now it will properly callback
in the event loop without issue.

This commit also addresses the lack of unit tests for
the parameter error issue -- there are now unit tests
for such.